requirements.txt
customtkinter==5.1.2
Pillow==8.4.0
tkVideo==0.1
moviepy==1.0.3
pydub==0.25.1